refactor: extract shared js utilities
