+++ src/types/html.rs
@@ -3,11 +3,11 @@ use std::path::Path;
 use vfs::VfsPath;
 
 use crate::LogLevel;
-use crate::types::js::{
+use crate::types::util::{
     JS_EXTENSIONS, is_node_builtin, resolve_alias_import, resolve_relative_import,
 };
 use crate::types::{Context, Edge, Parser};
-use crate::{Node, NodeKind, EdgeType};
+use crate::{EdgeType, Node, NodeKind};
 
 pub struct HtmlParser;
