+++ src/types/mod.rs
@@ -36,4 +36,5 @@ pub mod mdx;
 pub mod monorepo;
 pub mod package_json;
 pub mod package_util;
+pub mod util;
 pub mod vite;
